Add frontend code. and working on real time unread to read message update. 8th commit.

+++ 5.8.1/aspnet-core/src/RChat.Core/Messages/IMessageManager.cs
@@ -9,7 +9,7 @@ namespace RChat.Messages
     public interface IMessageManager : IDomainService
     {
         IEnumerable<Message> GetAllList();
-        IEnumerable<Message> GetAllListForUnReadMessages(int userPerRelationId);
+        IEnumerable<Message> GetAllListForUnReadMessages(long userPerRelationId);
         void UpdateUnReadMessageToRead(int userPerRelationId);
         IEnumerable<Message> GetAllForBothUser(int userPerRelationOne,int userPerRelationTwo);
         Task<Message> GetMessageById(int id);
